chore(build): uzfs build use submodule to build
