+++ Cargo.toml
@@ -6,37 +6,26 @@ edition = "2021"
 [lib]
 crate-type = ["staticlib", "rlib"]
 
-[workspace]
-members = ["uzfs-sys"]
-
 [build-dependencies]
 bindgen = "0.60"
-pkg-config = "0.3"
-uzfs-sys = { path = "uzfs-sys", version = "0.1.0" }
+pkg-config = "0.3.26"
 
-[workspace.dependencies]
-tokio-test = "0.4.2"
+[dependencies]
 tokio = { version = "1.32.0", features = ["macros", "rt-multi-thread", "signal", "time", "sync"] }
 futures = "0.3.28"
 dashmap = "5.4.0"
-serial_test = { version = "1.0.0" }
 once_cell = "1.13.1"
-rand = "0.8.5"
-
-[dependencies]
+libc = "0.2"
+backtrace = "0.3"
 cstr-argument = "0.1"
-uzfs-sys = { path = "uzfs-sys", version = "0.1.0" }
-tempfile = "3.3.0"
 prometheus-client = { git = "https://github.com/iomesh/prometheus_client_rust.git", tag = "v0.22.0-1"}
-tokio.workspace = true
-once_cell.workspace = true
 strum = "0.24"
 strum_macros = "0.24"
 
 [dev-dependencies]
-rand.workspace = true
+tokio-test = "0.4.2"
+rand = "0.8.5"
+serial_test = { version = "1.0.0" }
+tempfile = "3.3.0"
 petgraph = "0.6.3"
-serial_test.workspace = true
-dashmap.workspace = true
-tokio-test.workspace = true
 nix = { version = "0.27.1", features = ["process"] }
