+++ Makefile
@@ -1,4 +1,19 @@
-all: build
+top_srcdir=`pwd`
+scripts_dir=${top_srcdir}/scripts
+
+all: build_libuzfs_src build
+
+build_libuzfs_src:
+	@bash ${scripts_dir}/build_libuzfs_src.sh ${top_srcdir} ${ENABLE_DEBUG} ${ENABLE_ASAN}
+
+clean_libuzfs_src:
+	@bash ${scripts_dir}/clean_libuzfs_src.sh ${top_srcdir}
+
+build_uzfs_sys:
+	cargo build -p uzfs_sys
+
+build_uzfs:
+	cargo build -p uzfs
 
 build:
 	cargo build --all
@@ -9,9 +24,5 @@ test:
 	cargo test --target x86_64-unknown-linux-gnu uzfs_sync_test -- --ignored --nocapture
 	cargo test --target x86_64-unknown-linux-gnu --workspace  -- --nocapture
 
-clean:
+clean: clean_libuzfs_src
 	cargo clean
-	make -C uzfs-sys clean
-
-bench:
-	cargo bench --all
