--- uzfs-sys/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "uzfs-sys"
-version = "0.1.0"
-edition = "2021"
-
-[build-dependencies]
-bindgen = "0.60"
-pkg-config = "0.3.26"
-
-[dependencies]
-libc = "0.2"
-futures.workspace = true
-tokio.workspace = true
-dashmap.workspace = true
-once_cell.workspace = true
-backtrace = "0.3"
-
-[dev-dependencies]
-tokio-test.workspace = true
-rand.workspace = true
-
-[lib]
-crate-type = ["staticlib", "rlib"]
